Submit title input when focusing out (blur)
This patch ensures that the title input form is submitted when the focus on the
input is lost (`blur()` event).  The implementation adds a new `submitForm` method
on the Stimulus editable controller that is linked to an action on the input.
The form is submitted by using the `Rails.fire(form, event)` method form `@rails/ujs`.

todoapp specification:
- https://github.com/tastejs/todomvc/blob/master/app-spec.md#editing

Documentation:
- https://stimulusjs.org/handbook
- https://stimulusjs.org/reference
- https://developer.mozilla.org/en-US/docs/Web/API/Window/blur_event
- https://github.com/rails/rails/tree/master/actionview/app/assets/javascripts
- https://github.com/rails/rails/blob/master/actionview/app/assets/javascripts/rails-ujs/utils/event.coffee#L26-L34